+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - #79 Separate field initialization from curves (Thanks @alexander-zw)
 - #80 Add clang-tidy checks to library and CI
 - #82 Convert tests to use Google test (Thanks @alexander-zw)
+- #85 Add more unit tests for fields (Thanks @alexander-zw)
 ### Bug fixes
 - #75 Get rid of warning for unused constant PI, in complex field
 - #78 Reduce prints when inhibit_profiling_info is set